Evaluation_GenEval/requirements.txt
absl-py==1.4.0
addict==2.4.0
aiohttp==3.8.4
aiosignal==1.3.1
albumentations==1.3.0
altair==5.0.0
aniso8601==9.0.1
antlr4-python3-runtime==4.8
async-timeout==4.0.2
attrs==23.1.0
# autofaiss==2.15.8
blinker==1.6.2
braceexpand==0.1.7
cachetools==5.3.0
click==8.1.3
clip-anytorch==2.5.2
clip-benchmark==1.4.0
# clip-retrieval==2.37.0
cloudpickle==2.2.1
coloredlogs==15.0.1
contourpy==1.0.7
cycler==0.11.0
cython==0.29.34
dataclasses==0.6
decorator==5.1.1
docker-pycreds==0.4.0
einops==0.3.0
# embedding-reader==1.5.1
exifread-nocycle==3.0.1
faiss-cpu==1.7.4
filelock==3.12.0
fire==0.4.0
flask==2.3.3
flask-cors==3.0.10
flask-restful==0.3.10
flatbuffers==23.5.9
fonttools==4.39.4
frozenlist==1.3.3
# fsspec==2022.11.0
ftfy==6.1.1
future==0.18.3
gitdb==4.0.10
gitpython==3.1.31
google-auth==2.18.1
google-auth-oauthlib==1.0.0
grpcio==1.55.0
h5py==3.8.0
humanfriendly==10.0
imageio==2.9.0
imageio-ffmpeg==0.4.2
# img2dataset==1.42.0
importlib-metadata==6.6.0
importlib-resources==5.12.0
invisible-watermark==0.1.5
itsdangerous==2.1.2
jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
kiwisolver==1.4.4
kornia==0.6.0
lazy-loader==0.2
markdown==3.4.3
markdown-it-py==2.2.0
markupsafe==2.1.2
matplotlib==3.7.1
mdurl==0.1.2
mmcv-full==1.7.1
mmengine==0.7.3
model-index==0.1.11
mpmath==1.3.0
multidict==6.0.4
multilingual-clip==1.0.10
networkx==3.1
nltk==3.8.1
# nvidia-cublas-cu11==2022.4.8
# nvidia-cublas-cu117==11.10.1.25
# nvidia-cuda-runtime-cu11==2022.4.25
# nvidia-cuda-runtime-cu117==11.7.60
# nvidia-cudnn-cu11==2022.5.19
# nvidia-cudnn-cu116==8.4.0.27
# nvidia-cusolver-cu11==2022.4.8
# nvidia-cusolver-cu117==11.3.5.50
# nvidia-cusparse-cu11==2022.4.8
# nvidia-cusparse-cu117==11.7.3.50
# nvidia-pyindex==1.0.9
oauthlib==3.2.2
omegaconf==2.1.1
onnx==1.14.0
onnxruntime==1.14.1
open-clip-torch==2.20.0
opencv-python==4.6.0.66
opencv-python-headless==4.7.0.72
openmim==0.3.7
ordered-set==4.1.0
packaging==23.1
pandas==1.5.3
pathtools==0.1.2
prometheus-client==0.17.1
promise==2.3
protobuf==3.20.3
psutil==5.9.5
pyarrow==7.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycocoevalcap==1.2
pycocotools==2.0.6
pydeck==0.8.1b0
pydeprecate==0.3.1
pygments==2.15.1
pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
pytorch-lightning==1.4.2
pytz==2023.3
pywavelets==1.4.1
qudida==0.0.4
regex==2023.5.5
requests-oauthlib==1.3.1
rich==13.3.5
rsa==4.9
safetensors==0.3.1
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.9.1
semver==3.0.0
sentence-transformers==2.2.2
sentencepiece==0.1.99
sentry-sdk==1.29.2
setproctitle==1.3.2
shapely==2.0.1
shortuuid==1.0.11
smmap==5.0.0
streamlit==1.12.1
streamlit-drawable-canvas==0.8.0
submitit==1.4.5
sympy==1.12
tabulate==0.9.0
tensorboard==2.13.0
tensorboard-data-server==0.7.0
termcolor==2.3.0
terminaltables==3.1.10
test-tube==0.7.5
threadpoolctl==3.1.0
tifffile==2023.4.12
timm==0.9.2
tokenizers==0.15.0
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torchmetrics==0.6.0
tornado==6.3.2
tqdm==4.65.0
transformers==4.36.1
tzdata==2023.3
tzlocal==5.0.1
validators==0.20.0
wandb==0.12.21
watchdog==3.0.0
wcwidth==0.2.6
webdataset==0.2.48
werkzeug==2.3.7
yapf==0.33.0
yarl==1.9.2
zipp==3.15.0